Remove unnecessary apt dependencies for Package Manager.
Both libssl and sudo are already part of the base system, so we do not
need to install them directly. Moreover, it seems that libssl-dev is not
actually required to install and run RSPM, so we can remove it. Finally,
gdebi-core is only used during installation, so we can install it
temporarily during that step instead.

This has the following benefits:

* Fewer RUN steps.

* Fewer system packages in the resulting image, making it 10M smaller.

* Improved build time, because we don't have two apt-related steps (and
  their attendant cache cleanup code).

Signed-off-by: Aaron Jacobs <aaron.jacobs@rstudio.com>